@@ -1,9 +1,8 @@
-import React from 'react';
 import clsx from 'clsx';
+import Heading from '@theme/Heading';
 import styles from './styles.module.css';
 import Link from "@docusaurus/Link";
 
-// Keep the first and the 3rd empty to centerize the 2nd.
 const FeatureList = [
   {
   },
@@ -25,7 +24,7 @@ function Feature({Svg, title, description}) {
   return (
     <div className={clsx('col col--4')}>
       <div className="text--center padding-horiz--md">
-        <h3>{title}</h3>
+        <Heading as="h3">{title}</Heading>
         <p>{description}</p>
       </div>
     </div>
